fix: restructure evaluators into folders to match supported modalities
- [x] restructure evaluators into folders to match supported modalities

Will be done in a future PR:

- [ ] rename evaluators to reflect supported modalities. I would probably drop `Any` and instead rename old ones to `Text*`
- [ ] make evaluators a private module

related to  #1337
previously PR on this #2980
