+++ mteb/evaluation/evaluators/__init__.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from .any_sts_evaluator import AnySTSEvaluator
-from .bitext_mining_evaluator import BitextMiningEvaluator
+from .text.bitext_mining_evaluator import BitextMiningEvaluator
 from .classification_evaluator import ClassificationEvaluator
 from .clustering_evaluator import ClusteringEvaluator
 from .evaluator import Evaluator
@@ -10,10 +10,10 @@ from .image import (
     Any2AnyRetrievalEvaluator,
     ImageTextPairClassificationEvaluator,
 )
-from .pair_classification_evaluator import PairClassificationEvaluator
-from .RegressionEvaluator import LinearRegressionEvaluator
+from .text.pair_classification_evaluator import PairClassificationEvaluator
+from .image.regression_evaluator import LinearRegressionEvaluator
 from .retrieval_evaluator import RetrievalEvaluator
-from .summarization_evaluator import (
+from .text.summarization_evaluator import (
     DeprecatedSummarizationEvaluator,
     SummarizationEvaluator,
 )
