+++ mteb/evaluation/evaluators/text/bitext_mining_evaluator.py
@@ -12,8 +12,8 @@ from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_sc
 from mteb.abstasks.task_metadata import TaskMetadata
 from mteb.models.models_protocols import Encoder
 
-from ...create_dataloaders import create_dataloader_from_texts
-from .evaluator import Evaluator
+from ....create_dataloaders import create_dataloader_from_texts
+from ..evaluator import Evaluator
 
 logger = logging.getLogger(__name__)
