+++ mteb/evaluation/evaluators/text/pair_classification_evaluator.py
@@ -16,9 +16,9 @@ from mteb.abstasks.task_metadata import TaskMetadata
 from mteb.models.model_meta import ScoringFunction
 from mteb.models.models_protocols import Encoder
 
-from ...create_dataloaders import create_dataloader_from_texts
-from ...similarity_functions import compute_pairwise_similarity
-from .evaluator import Evaluator
+from ....create_dataloaders import create_dataloader_from_texts
+from ....similarity_functions import compute_pairwise_similarity
+from ..evaluator import Evaluator
 
 logger = logging.getLogger(__name__)
