+++ mteb/evaluation/evaluators/text/summarization_evaluator.py
@@ -13,8 +13,8 @@ from mteb.abstasks.task_metadata import TaskMetadata
 from mteb.models.models_protocols import Encoder
 from mteb.similarity_functions import cos_sim, dot_score
 
-from ...create_dataloaders import create_dataloader_from_texts
-from .evaluator import Evaluator
+from ....create_dataloaders import create_dataloader_from_texts
+from ..evaluator import Evaluator
 
 # if later than python 3.13 use typing module
 if sys.version_info >= (3, 13):
